pyproject.toml
[project]
name = "agente-whind"
version = "0.1.0"
description = "Sistema multiagente PM + Tech Writer para geração de PRDs via Telegram"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.14",
    "openai>=2.13.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "PyGithub>=2.0.0",
    "sqlalchemy>=2.0.0",
]
